+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Client Versions ##
+* v9.40.0 (2014-08-19)
+	* allow dividends to be paid to XCP holders (protocol change: 317000)
+	* fixed bug in BTCpay validation
+	* allow null expirations
+	* assert first block in database is BLOCK_FIRST
 * v9.39.0 (2014-08-06)
 	* re‐match expired order matches from a new block all at once (protocol change: 315000)
 	* bug in issuance fee (protocol change: 315000)